update on 'update_expression' function. Need to add special conditions to deal with different cases or maybe i''l find another way

+++ ExS.hpp
@@ -25,8 +25,8 @@ namespace ExS
     bool is_a_minus_or_dot(const char &character);
     void loop(std::string &expr);
     size_t choose_operator(const ExS::IntCharMap &operator_map);
-    std::tuple<float, float> bi_operator_operands(std::string &expr, size_t &priority_index, IntCharMap &operator_pos_map);
-    std::string update_expression(std::string &expr, float &result, size_t sub_expr_begin_index, size_t sub_expr_end_index);
+    std::tuple<double, double> bi_operator_operands(std::string &expr, size_t &priority_index, IntCharMap &operator_pos_map);
+    std::string update_expression(std::string &expr, double &result, size_t sub_expr_begin_index, size_t sub_expr_end_index);
     IntCharMap verify_and_collect_pos(const std::string &expr);
 }
 #endif
